+++ Kernel/Config/Defaults.pm
@@ -423,6 +423,22 @@ sub LoadDefaults {
 #    $Self->{'AuthModule::Radius::Die'} = 1;
 
     # --------------------------------------------------- #
+    # 2 factor authentication settings                    #
+    # check a otp (one-time password)                     #
+    # after successful authentication                     #
+    # as an extra security measure                        #
+    # --------------------------------------------------- #
+    # This is the auth module using the google authenticator mechanism
+#    $Self->{'AuthTwoFactorModule'} = 'Kernel::System::Auth::TwoFactor::GoogleAuthenticator';
+
+    # defines user preference where the secret key is stored
+#    $Self->{'AuthTwoFactorModule::SecretPreferencesKey'} = 'GoogleAuthenticatorSecretKey';
+
+    # defines if the otp for the previous timespan (30-60sec ago) will also be valid
+    # helpful to account for timimg issues (server and entry based)
+#    $Self->{'AuthTwoFactorModule::AllowPreviousToken'} = '1';
+
+    # --------------------------------------------------- #
     # authentication sync settings                        #
     # (enable agent data sync. after succsessful          #
     # authentication)                                     #
@@ -1304,6 +1320,22 @@ via the Preferences button after logging in.
 #    $Self->{'Customer::AuthModule'} = 'Kernel::System::Auth::Radius';
 #    $Self->{'Customer::AuthModule::Radius::Host'} = 'radiushost';
 #    $Self->{'Customer::AuthModule::Radius::Password'} = 'radiussecret';
+
+    # --------------------------------------------------- #
+    # 2 factor customer authentication settings           #
+    # check a otp (one-time password)                     #
+    # after successful authentication                     #
+    # as an extra security measure                        #
+    # --------------------------------------------------- #
+    # This is the auth module using the google authenticator mechanism
+#    $Self->{'Customer::AuthTwoFactorModule'} = 'Kernel::System::CustomerAuth::TwoFactor::GoogleAuthenticator';
+
+    # defines user preference where the secret key is stored
+#    $Self->{'Customer::AuthTwoFactorModule::SecretPreferencesKey'} = 'GoogleAuthenticatorSecretKey';
+
+    # defines if the otp for the previous timespan (30-60sec ago) will also be valid
+    # helpful to account for timimg issues (server and entry based)
+#    $Self->{'Customer::AuthTwoFactorModule::AllowPreviousToken'} = '1';
 
     # --------------------------------------------------- #
     #                                                     #