+++ Kernel/System/Auth.pm
@@ -72,6 +72,21 @@ sub new {
         $Self->{"AuthBackend$Count"} = $GenericModule->new( Count => $Count );
     }
 
+    # load 2factor auth modules
+    COUNT:
+    for my $Count ( '', 1 .. 10 ) {
+
+        my $GenericModule = $ConfigObject->Get("AuthTwoFactorModule$Count");
+
+        next COUNT if !$GenericModule;
+
+        if ( !$MainObject->Require($GenericModule) ) {
+            $MainObject->Die("Can't load backend module $GenericModule! $@");
+        }
+
+        $Self->{"AuthTwoFactorBackend$Count"} = $GenericModule->new( %{$Self}, Count => $Count );
+    }
+
     # load sync modules
     COUNT:
     for my $Count ( '', 1 .. 10 ) {
@@ -143,6 +158,36 @@ sub Auth {
         # next on no success
         next COUNT if !$User;
 
+        my $UserID = $UserObject->UserLookup(
+            UserLogin => $User,
+        );
+
+        # check 2factor auth backends
+        my $TwoFactorAuth;
+        TWOFACTORSOURCE:
+        for my $Count ( '', 1 .. 10 ) {
+
+            # return on no config setting
+            next TWOFACTORSOURCE if !$Self->{"AuthTwoFactorBackend$Count"};
+
+            # 2factor backend
+            my $AuthOk = $Self->{"AuthTwoFactorBackend$Count"}->Auth(
+                TwoFactorToken => $Param{TwoFactorToken},
+                User => $User,
+                UserID => $UserID,
+            );
+            $TwoFactorAuth = $AuthOk ? 'passed' : 'failed';
+
+            last TWOFACTORSOURCE if $AuthOk;
+        }
+
+        # if at least one 2factor auth backend was checked but none was successful,
+        # it counts as a failed login
+        if ( $TwoFactorAuth && $TwoFactorAuth ne 'passed' ) {
+            $User = undef;
+            last COUNT;
+        }
+
         # configured auth sync backend
         my $AuthSyncBackend = $ConfigObject->Get("AuthModule::UseSyncBackend$Count");
         if ( !defined $AuthSyncBackend ) {
@@ -176,10 +221,6 @@ sub Auth {
         }
 
         # remember auth backend
-        my $UserID = $UserObject->UserLookup(
-            UserLogin => $User,
-        );
-
         if ($UserID) {
             $UserObject->SetPreferences(
                 Key    => 'UserAuthBackend',
@@ -188,7 +229,7 @@ sub Auth {
             );
         }
 
-        last COUNT if $User;
+        last COUNT;
     }
 
     # return if no auth user