+++ Kernel/System/Web/InterfaceCustomer.pm
@@ -218,6 +218,10 @@ sub Run {
             Param => 'Password',
             Raw   => 1
         ) || '';
+        my $PostTwoFactorToken = $ParamObject->GetParam(
+            Param => 'TwoFactorToken',
+            Raw   => 1
+        ) || '';
 
         # create AuthObject
         my $AuthObject = $Kernel::OM->Get('Kernel::System::CustomerAuth');
@@ -225,7 +229,8 @@ sub Run {
         # check submitted data
         my $User = $AuthObject->Auth(
             User => $PostUser,
-            Pw   => $PostPw
+            Pw   => $PostPw,
+            TwoFactorToken => $PostTwoFactorToken,
         );
 
         my $Expires = '+' . $ConfigObject->Get('SessionMaxTime') . 's';